backend/requirements.txt
fastapi
uvicorn
torch
torchvision
torchaudio
transformers
sentence-transformers
numpy
scipy
scikit-learn
pandas
requests
beautifulsoup4
arxiv
neurokit2
brian2
mediapipe
opencv-python-headless
# pyaudio  # Optional: Requires system audio libraries (PortAudio)
#         # Install system dependencies first:
#         # macOS: brew install portaudio
#         # Ubuntu/Debian: sudo apt-get install portaudio19-dev
#         # Then: pip install pyaudio
websockets
pyyaml
python-multipart
python-jose[cryptography]
passlib[bcrypt]
# Fix pyasn1 version conflict - use versions compatible with python-jose
pyasn1>=0.5.0
pyasn1-modules>=0.3.0
# Quantum Computing Dependencies
dimod
# dwave-networkx  # Optional: specific version issues on some OS
# dwave-systems   # Optional: specific version issues on some OS
# dwave-ocean-sdk # Alternative if above fail
pennylane
qiskit
qiskit-algorithms
qiskit-optimization
qiskit-aer
networkx
# Multi-Agent System - Swarms Framework for Agentic LLM Operations
# Reference: https://github.com/kyegomez/swarms.git
swarms>=2.0.0
# Database & Storage
psycopg2-binary
supabase
PyPDF2
python-docx
# Lemon AI (optional - can be installed separately)

# Swarms Tools - Extended agent capabilities
# Install from: https://github.com/The-Swarm-Corporation/swarms-tools.git
swarms-tools>=1.0.0

# SwarmDB - Enhanced database functionality
# Install from: https://github.com/The-Swarm-Corporation/SwarmDB.git
# Note: May need to install from source
# git+https://github.com/The-Swarm-Corporation/SwarmDB.git

# Additional dependencies for swarms-tools
httpx>=0.24.0
orjson>=3.9.0

# Note: To install SwarmDB, clone and install from source:
# git clone https://github.com/The-Swarm-Corporation/SwarmDB.git
# cd SwarmDB
# pip install -e .

# Swarm Corporation Components
# Security
swarm-shield>=0.0.1

# Document Processing
doc-master>=1.0.0
omniparse>=1.0.0
agentparse>=1.0.0

# Research
research-paper-hive>=1.0.0
advanced-research>=1.0.0

# RAG
agent-rag-protocol>=1.0.0
multi-agent-rag-template>=1.0.0

# Workflow
zero>=1.0.0

# Utilities
swarms-utils>=1.0.0

# Note: Many components may need to be installed from source:
# git clone https://github.com/The-Swarm-Corporation/<component>.git
# cd <component>
# pip install -e .

# Additional dependencies
pyyaml>=6.0
